+++ crates/server/src/main.rs
@@ -3,6 +3,7 @@ mod app_state;
 mod handlers;
 mod services;
 mod telegram_bot_service;
+mod scarb_binaries_manager_service;
 
 use app_state::AppState;
 use aws_config::meta::region::RegionProviderChain;
@@ -37,7 +38,11 @@ use std::env::consts::ARCH;
 use std::fs;
 use std::os::unix::fs::PermissionsExt;
 use std::path::Path;
-
+use clokwerk::{Job, AsyncScheduler, TimeUnits};
+use tokio::spawn;
+use tokio::time::{interval, Duration};
+use tracing::{error, info};
+use crate::scarb_binaries_manager_service::{download_custom_scarb_binaries, start_github_scarb_binaries_downloader_scheduler};
 // Resources
 // https://github.com/tokio-rs/axum/tree/main/examples
 // https://www.apianalytics.dev/
@@ -81,13 +86,9 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let s3_client = Client::new(&shared_config);
     sqlx::migrate!().run(&db_pool).await?;
 
-    download_binary(&s3_client, format!("sozo/{ARCH}/sozo_v1.0.1").as_str()).await?;
-    download_binary(&s3_client, format!("scarb/{ARCH}/scarb_cairo_v_2_6_3").as_str()).await?;
-    download_binary(&s3_client, format!("scarb/{ARCH}/scarb_cairo_v_2_6_4").as_str()).await?;
-    download_binary(&s3_client, format!("scarb/{ARCH}/scarb_cairo_v_2_7_0").as_str()).await?;
-    download_binary(&s3_client, format!("scarb/{ARCH}/scarb_cairo_v2.8.2").as_str()).await?;
-    download_binary(&s3_client, format!("scarb/{ARCH}/scarb_cairo_v2.8.4").as_str()).await?;
-    download_binary(&s3_client, format!("scarb/{ARCH}/scarb_cairo_v2.8.5").as_str()).await?;
+    // Download scarb binaries
+    download_custom_scarb_binaries(&s3_client).await?;
+    start_github_scarb_binaries_downloader_scheduler().await;
 
     let shared_state = Arc::new(AppState {
         db_pool,
@@ -133,52 +134,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .serve(app.into_make_service())
         .await
         .unwrap();
-
-    Ok(())
-}
-
-// downloads the binary from the S3 bucket and saves it to the local directory, gives the executable permissions
-async fn download_binary(
-    s3_client: &Client,
-    object_key: &str,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let bucket_name = std::env::var("BINARIES_S3_BUCKET_NAME").unwrap_or("".to_string());
-    let binaries_save_directory_path = std::env::var("BINARIES_SAVE_DIRECTORY_PATH").unwrap_or("".to_string());
-    let local_file_path = format!("{}/{}", &binaries_save_directory_path, &object_key.replace(format!("/{}/", ARCH).as_str(), "/"));
-    // Check if the file already exists
-    let path = Path::new(&local_file_path);
-    if path.exists() {
-        println!("File already exists (skipping download): {}", local_file_path);
-        return Ok(()); // Exit early if the file exists
-    }
-    println!("Downloading object: {}/{}", bucket_name, object_key);
-
-    // Fetch the object from the S3 bucket
-    let resp = s3_client
-        .get_object()
-        .bucket(bucket_name)
-        .key(object_key)
-        .send()
-        .await?;
-
-    // Ensure the directory exists
-    if let Some(parent_dir) = std::path::Path::new(&local_file_path).parent() {
-        fs::create_dir_all(parent_dir).expect("Failed to create parent directories");
-    }
-    let mut file = File::create(&local_file_path).expect(format!("Failed to create file: {}", local_file_path).as_str());
-
-    // Stream the object content to the file
-    let data = resp.body.collect().await?;
-    file.write_all(&data.into_bytes())
-        .expect("Failed to write object data to file");
-
-    let metadata = fs::metadata(&local_file_path)?;
-    let mut permissions = metadata.permissions();
-    permissions.set_mode(0o755); // rwxr-xr-x
-    fs::set_permissions(&local_file_path, permissions)
-        .expect("Failed to set executable permissions");
-
-    println!("Object downloaded successfully to: {}", local_file_path);
 
     Ok(())
 }